+++ bridges/otelslog/handler.go
@@ -201,10 +201,10 @@ func (h *Handler) convertRecord(r slog.Record) log.Record {
 // Enable returns true if the Handler is enabled to log for the provided
 // context and Level. Otherwise, false is returned if it is not enabled.
 func (h *Handler) Enabled(ctx context.Context, l slog.Level) bool {
-	var record log.Record
+	var param log.EnabledParameters
 	const sevOffset = slog.Level(log.SeverityDebug) - slog.LevelDebug
-	record.SetSeverity(log.Severity(l + sevOffset))
-	return h.logger.Enabled(ctx, record)
+	param.SetSeverity(log.Severity(l + sevOffset))
+	return h.logger.Enabled(ctx, param)
 }
 
 // WithAttrs returns a new [slog.Handler] based on h that will log using the
@@ -425,7 +425,12 @@ func convertValue(v slog.Value) log.Value {
 	case slog.KindTime:
 		return log.Int64Value(v.Time().UnixNano())
 	case slog.KindUint64:
-		return log.Int64Value(int64(v.Uint64()))
+		const maxInt64 = ^uint64(0) >> 1
+		u := v.Uint64()
+		if u > maxInt64 {
+			return log.Float64Value(float64(u))
+		}
+		return log.Int64Value(int64(u)) // nolint:gosec  // Overflow checked above.
 	case slog.KindGroup:
 		g := v.Group()
 		buf := newKVBuffer(len(g))