+++ processors/minsev/minsev_test.go
@@ -22,27 +22,36 @@ var severities = []api.Severity{
 	api.SeverityFatal, api.SeverityFatal1, api.SeverityFatal2, api.SeverityFatal3, api.SeverityFatal4,
 }
 
-type args struct {
+type apiSev api.Severity
+
+func (s apiSev) Severity() api.Severity { return api.Severity(s) }
+
+type emitArgs struct {
 	Ctx    context.Context
-	Record log.Record
+	Record *log.Record
+}
+
+type enabledArgs struct {
+	Ctx   context.Context
+	Param api.EnabledParameters
 }
 
 type processor struct {
 	ReturnErr error
 
-	OnEmitCalls     []args
-	EnabledCalls    []args
+	OnEmitCalls     []emitArgs
+	EnabledCalls    []enabledArgs
 	ForceFlushCalls []context.Context
 	ShutdownCalls   []context.Context
 }
 
-func (p *processor) OnEmit(ctx context.Context, r log.Record) error {
-	p.OnEmitCalls = append(p.OnEmitCalls, args{ctx, r})
+func (p *processor) OnEmit(ctx context.Context, r *log.Record) error {
+	p.OnEmitCalls = append(p.OnEmitCalls, emitArgs{ctx, r})
 	return p.ReturnErr
 }
 
-func (p *processor) Enabled(ctx context.Context, r log.Record) bool {
-	p.EnabledCalls = append(p.EnabledCalls, args{ctx, r})
+func (p *processor) Enabled(ctx context.Context, param api.EnabledParameters) bool {
+	p.EnabledCalls = append(p.EnabledCalls, enabledArgs{ctx, param})
 	return true
 }
 
@@ -63,20 +72,42 @@ func (p *processor) Reset() {
 	p.ForceFlushCalls = p.ForceFlushCalls[:0]
 }
 
+func TestLogProcessorDynamicSeverity(t *testing.T) {
+	sev := new(SeverityVar)
+	wrapped := new(processor)
+	p := NewLogProcessor(wrapped, sev)
+
+	ctx := context.Background()
+	params := &api.EnabledParameters{}
+	params.SetSeverity(api.SeverityDebug)
+	assert.False(t, p.Enabled(ctx, *params), api.SeverityDebug.String())
+
+	params.SetSeverity(api.SeverityInfo)
+	assert.True(t, p.Enabled(ctx, *params), api.SeverityInfo.String())
+
+	sev.Set(SeverityError)
+
+	params.SetSeverity(api.SeverityInfo)
+	assert.False(t, p.Enabled(ctx, *params), api.SeverityInfo.String())
+
+	params.SetSeverity(api.SeverityError)
+	assert.True(t, p.Enabled(ctx, *params), api.SeverityError.String())
+}
+
 func TestLogProcessorOnEmit(t *testing.T) {
 	t.Run("Passthrough", func(t *testing.T) {
 		wrapped := &processor{ReturnErr: assert.AnError}
 
-		p := NewLogProcessor(wrapped, api.SeverityTrace1)
+		p := NewLogProcessor(wrapped, SeverityTrace1)
 		ctx := context.Background()
 		r := &log.Record{}
 		for _, sev := range severities {
 			r.SetSeverity(sev)
-			assert.ErrorIs(t, p.OnEmit(ctx, *r), assert.AnError, sev.String())
+			assert.ErrorIs(t, p.OnEmit(ctx, r), assert.AnError, sev.String())
 
 			if assert.Lenf(t, wrapped.OnEmitCalls, 1, "Record with severity %s not passed-through", sev) {
 				assert.Equal(t, ctx, wrapped.OnEmitCalls[0].Ctx, sev.String())
-				assert.Equal(t, *r, wrapped.OnEmitCalls[0].Record, sev.String())
+				assert.Equal(t, r, wrapped.OnEmitCalls[0].Record, sev.String())
 			}
 			wrapped.Reset()
 		}
@@ -85,14 +116,14 @@ func TestLogProcessorOnEmit(t *testing.T) {
 	t.Run("Dropped", func(t *testing.T) {
 		wrapped := &processor{ReturnErr: assert.AnError}
 
-		p := NewLogProcessor(wrapped, api.SeverityFatal4+1)
+		p := NewLogProcessor(wrapped, apiSev(api.SeverityFatal4+1))
 		ctx := context.Background()
 		r := &log.Record{}
 		for _, sev := range severities {
 			r.SetSeverity(sev)
-			assert.NoError(t, p.OnEmit(ctx, *r), assert.AnError, sev.String())
+			assert.NoError(t, p.OnEmit(ctx, r), sev.String())
 
-			if !assert.Lenf(t, wrapped.OnEmitCalls, 0, "Record with severity %s passed-through", sev) {
+			if !assert.Emptyf(t, wrapped.OnEmitCalls, "Record with severity %s passed-through", sev) {
 				wrapped.Reset()
 			}
 		}
@@ -103,16 +134,16 @@ func TestLogProcessorEnabled(t *testing.T) {
 	t.Run("Passthrough", func(t *testing.T) {
 		wrapped := &processor{}
 
-		p := NewLogProcessor(wrapped, api.SeverityTrace1)
+		p := NewLogProcessor(wrapped, SeverityTrace1)
 		ctx := context.Background()
-		r := &log.Record{}
+		param := api.EnabledParameters{}
 		for _, sev := range severities {
-			r.SetSeverity(sev)
-			assert.True(t, p.Enabled(ctx, *r), sev.String())
+			param.SetSeverity(sev)
+			assert.True(t, p.Enabled(ctx, param), sev.String())
 
 			if assert.Lenf(t, wrapped.EnabledCalls, 1, "Record with severity %s not passed-through", sev) {
 				assert.Equal(t, ctx, wrapped.EnabledCalls[0].Ctx, sev.String())
-				assert.Equal(t, *r, wrapped.EnabledCalls[0].Record, sev.String())
+				assert.Equal(t, param, wrapped.EnabledCalls[0].Param, sev.String())
 			}
 			wrapped.Reset()
 		}
@@ -121,24 +152,44 @@ func TestLogProcessorEnabled(t *testing.T) {
 	t.Run("NotEnabled", func(t *testing.T) {
 		wrapped := &processor{}
 
-		p := NewLogProcessor(wrapped, api.SeverityFatal4+1)
+		p := NewLogProcessor(wrapped, apiSev(api.SeverityFatal4+1))
 		ctx := context.Background()
-		r := &log.Record{}
+		param := api.EnabledParameters{}
 		for _, sev := range severities {
-			r.SetSeverity(sev)
-			assert.False(t, p.Enabled(ctx, *r), sev.String())
+			param.SetSeverity(sev)
+			assert.False(t, p.Enabled(ctx, param), sev.String())
 
-			if !assert.Lenf(t, wrapped.EnabledCalls, 0, "Record with severity %s passed-through", sev) {
+			if !assert.Emptyf(t, wrapped.EnabledCalls, "Record with severity %s passed-through", sev) {
 				wrapped.Reset()
 			}
 		}
+	})
+
+	t.Run("NoFiltered", func(t *testing.T) {
+		wrapped := &processor{}
+
+		pruned := struct{ log.Processor }{wrapped} // Remove the Enabled method.
+		p := NewLogProcessor(pruned, SeverityInfo)
+		ctx := context.Background()
+		params := &api.EnabledParameters{}
+
+		params.SetSeverity(api.SeverityDebug)
+		assert.False(t, p.Enabled(ctx, *params))
+
+		params.SetSeverity(api.SeverityInfo)
+		assert.True(t, p.Enabled(ctx, *params))
+
+		params.SetSeverity(api.SeverityError)
+		assert.True(t, p.Enabled(ctx, *params))
+
+		assert.Empty(t, wrapped.EnabledCalls)
 	})
 }
 
 func TestLogProcessorForceFlushPassthrough(t *testing.T) {
 	wrapped := &processor{ReturnErr: assert.AnError}
 
-	p := NewLogProcessor(wrapped, api.SeverityTrace1)
+	p := NewLogProcessor(wrapped, SeverityTrace1)
 	ctx := context.Background()
 	assert.ErrorIs(t, p.ForceFlush(ctx), assert.AnError)
 	assert.Len(t, wrapped.ForceFlushCalls, 1, "ForceFlush not passed-through")
@@ -147,37 +198,51 @@ func TestLogProcessorForceFlushPassthrough(t *testing.T) {
 func TestLogProcessorShutdownPassthrough(t *testing.T) {
 	wrapped := &processor{ReturnErr: assert.AnError}
 
-	p := NewLogProcessor(wrapped, api.SeverityTrace1)
+	p := NewLogProcessor(wrapped, SeverityTrace1)
 	ctx := context.Background()
 	assert.ErrorIs(t, p.Shutdown(ctx), assert.AnError)
 	assert.Len(t, wrapped.ShutdownCalls, 1, "Shutdown not passed-through")
 }
 
+func TestLogProcessorNilSeverity(t *testing.T) {
+	p := NewLogProcessor(nil, nil)
+	assert.Equal(t, SeverityInfo, p.sev.(Severity))
+}
+
 func TestLogProcessorNilDownstream(t *testing.T) {
-	p := NewLogProcessor(nil, api.SeverityTrace1)
+	p := NewLogProcessor(nil, SeverityTrace1)
 	ctx := context.Background()
-	r := log.Record{}
+	r := new(log.Record)
 	r.SetSeverity(api.SeverityTrace1)
+	param := api.EnabledParameters{}
+	param.SetSeverity(api.SeverityTrace1)
 	assert.NotPanics(t, func() {
 		assert.NoError(t, p.OnEmit(ctx, r))
-		assert.False(t, p.Enabled(ctx, r))
+		assert.False(t, p.Enabled(ctx, param))
 		assert.NoError(t, p.ForceFlush(ctx))
 		assert.NoError(t, p.Shutdown(ctx))
 	})
 }
 
 func BenchmarkLogProcessor(b *testing.B) {
-	rPtr := new(log.Record)
-	rPtr.SetSeverity(api.SeverityTrace)
-	ctx, r := context.Background(), *rPtr
+	r := new(log.Record)
+	r.SetSeverity(api.SeverityTrace)
+	param := api.EnabledParameters{}
+	param.SetSeverity(api.SeverityTrace)
+	ctx := context.Background()
 
-	run := func(p log.Processor) func(b *testing.B) {
+	type combo interface {
+		log.Processor
+		filterProcessor
+	}
+
+	run := func(p combo) func(b *testing.B) {
 		return func(b *testing.B) {
 			var err error
 			var enabled bool
 			b.ReportAllocs()
 			for n := 0; n < b.N; n++ {
-				enabled = p.Enabled(ctx, r)
+				enabled = p.Enabled(ctx, param)
 				err = p.OnEmit(ctx, r)
 			}
 
@@ -186,6 +251,6 @@ func BenchmarkLogProcessor(b *testing.B) {
 	}
 
 	b.Run("Base", run(defaultProcessor))
-	b.Run("Enabled", run(NewLogProcessor(nil, api.SeverityTrace)))
-	b.Run("Disabled", run(NewLogProcessor(nil, api.SeverityDebug)))
+	b.Run("Enabled", run(NewLogProcessor(nil, SeverityTrace)))
+	b.Run("Disabled", run(NewLogProcessor(nil, SeverityDebug)))
 }